Use floats instead of doubles. Create FloatMatrix class, a 2D matrix wrapper, to resolve memory issues of the HungarianAlgorithm when used with large datasets.

+++ src/main/java/mp_pprl/core/Party.java
@@ -156,7 +156,7 @@ public class Party {
         }
     }
 
-    public void generateNoise(double noisePercentage) {
+    public void generateNoise(float noisePercentage) {
         List<List<String>> noiseData = NoiseDataGenerator.generateNoiseData(soundexEncodedRecords.size(), noisePercentage);
         soundexEncodedRecords.addAll(noiseData);
     }
